requirements.txt
Django>=5.2.5
djangorestframework>=3.14.0 